restart
redoing my css to make for a cleaner set up. Commented out large sections of code to get back to basics. Re did the nav bar, and floating contact. Need to work on containers and setting up a side well for the contact bar to float in. Maybe move it to the right?